Fix prefix escaping at ShardClient.keys function
Currently, the keys function uses the make_key function which doesn't
escape the prefix. In case the default * prefix is used and some
other client uses the same redis db with non-default prefix, it
misleads to fetching of prefixed keys of the other client.
In order to fix the issue, we have to use the make_pattern function
instead of the make_key function.